Swapped combineChunks
